requirements.txt
librosa==0.9.2
numpy==1.24.3
resampy==0.4.2
scipy==1.11.3
pysoundfile==0.12.1
pytorch==2.1.1